chore: rename description folder to generic description
