--- src/Common/Description/constant.ts
@@ -1,3 +1,0 @@
-const Moment12HourFormat = 'ddd, DD MMM YYYY, hh:mm A'
-
-export default Moment12HourFormat
