+++ src/Common/GenericDescription/constant.ts
@@ -1,0 +1,3 @@
+export const Moment12HourFormat = 'ddd, DD MMM YYYY, hh:mm A'
+
+export const README = 'Readme'
