+++ src/Common/GenericDescription/genericDescription.scss
