+++ src/Common/index.ts
@@ -51,4 +51,4 @@ export * from './CustomInput'
 export * from './DraggableWrapper'
 export * from './Pagination'
 export * from './Markdown'
-export * from './Description'
+export * from './GenericDescription'
